+++ UAParser.Tests/YamlTestCase.cs
@@ -2,27 +2,28 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using UAParser.Implementations;
 using Xunit;
 
 namespace UAParser.Tests
 {
-  public abstract class YamlTestCase
-  {
-    public string UserAgent { get; set; }
-    public abstract void Verify(ClientInfo clientInfo);
-
-    protected void AssertMatch<T>(T expected, T actual, string type)
+    public abstract class YamlTestCase
     {
-      if (typeof(T) == typeof(string))
-      {
-        string exp = expected as string;
-        string act = actual as string;
+        public string UserAgent { get; set; }
+        public abstract void Verify(ClientInfo clientInfo);
 
-        if (string.IsNullOrEmpty(exp) && string.IsNullOrEmpty(act))
-          return;
-      }
+        protected void AssertMatch<T>(T expected, T actual, string type)
+        {
+            if (typeof(T) == typeof(string))
+            {
+                string exp = expected as string;
+                string act = actual as string;
 
-      Assert.True(expected.Equals(actual), type+" did not match. (expected:" + expected + " actual:" + actual + ")  in " + UserAgent);
+                if (string.IsNullOrEmpty(exp) && string.IsNullOrEmpty(act))
+                    return;
+            }
+
+            Assert.True(expected.Equals(actual), type + " did not match. (expected:" + expected + " actual:" + actual + ")  in " + UserAgent);
+        }
     }
-  }
 }
