+++ UAParser/VersionString.cs
@@ -1,0 +1,34 @@
+#region Apache License, Version 2.0
+//
+// Copyright 2014 Atif Aziz
+// Portions Copyright 2012 Søren Enemærke
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+#endregion
+
+
+namespace UAParser
+{
+    using System;
+    using System.Linq;
+
+    internal static class VersionString
+    {
+        public static string Format(params string[] parts)
+        {
+            return string.Join(".", parts.Where(v => !String.IsNullOrEmpty(v)).ToArray());
+        }
+    }
+
+}
